Move temporal models into scaling module
