+++ src/prfmodel/_docstring.py
@@ -66,10 +66,10 @@ _PARAMS: dict[str, str] = {
         "    instance with default values."
     ),
     "model_prf": ("prf_model : BaseResponse\n    A population receptive field response model instance."),
-    "model_temporal": (
-        "temporal_model : BaseTemporal or type or None, default=BaselineAmplitude, optional\n"
+    "model_scaling": (
+        "scaling_model : BaseScaling or type or None, default=BaselineAmplitude, optional\n"
         "    A temporal model class or instance. Model classes will be instantiated during initialization.\n"
-        "    The default creates a :class:`~prfmodel.models.temporal.BaselineAmplitude` instance."
+        "    The default creates a :class:`~prfmodel.models.scaling.BaselineAmplitude` instance."
     ),
     "parameters": (
         "parameters : pandas.DataFrame\n"
